Update action names in CI workflows
Arduino's CI actions have been moved to dedicated repositories. The copies that remain in the old repository are no
longer maintained. This requires an update to the action names in the workflows.

The default branch name in the new repositories is `main`, as opposed to the `master` of the previous repo. This
requires the action refs to be updated as well.

The default sketches reports folder name/workflow artifact name was changed, which requires updating the
actions/upload-artifact step. In order to avoid further reliance on the default value, I have explicitely defined the
name in the workflows.