Use modern API of the arduino/compile-sketches CI action
The size-report-sketch input is no longer used. Reports are made for all compiled sketches.

The enable-size-deltas-report input was renamed enable-deltas-report.

Because these changes were made in a backwards compatible manner, the previous use of the outdated API worked fine, but
it did result in warnings in the workflow run log, which might cause confusion.